Refactor: extract common pto_runtime_c_api functions into shared source
The onboard and sim versions of pto_runtime_c_api.cpp were ~250 lines each
with 11 of 13 functions identical. Extract the shared functions into
platform/src/host/pto_runtime_c_api_common.cpp, leaving only the 2
platform-specific functions (launch_runtime, set_device) in each variant.
